azure-intellij-plugin-storage
