+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-storage/src/main/java/com/microsoft/azure/toolkit/intellij/storage/IntellijStorageActionsContributor.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License. See License.txt in the project root for license information.
+ */
+
+package com.microsoft.azure.toolkit.intellij.storage;
+
+import com.intellij.openapi.actionSystem.AnActionEvent;
+import com.microsoft.azure.toolkit.ide.common.IActionsContributor;
+import com.microsoft.azure.toolkit.ide.common.action.AzureActionManager;
+import com.microsoft.azure.toolkit.ide.common.action.ResourceCommonActionsContributor;
+import com.microsoft.azure.toolkit.intellij.storage.creation.CreateStorageAccountAction;
+import com.microsoft.azure.toolkit.lib.storage.service.AzureStorageAccount;
+
+import java.util.function.BiConsumer;
+import java.util.function.BiPredicate;
+
+public class IntellijStorageActionsContributor implements IActionsContributor {
+    @Override
+    public void registerHandlers(AzureActionManager am) {
+        final BiPredicate<Object, AnActionEvent> condition = (r, e) -> r instanceof AzureStorageAccount;
+        final BiConsumer<Object, AnActionEvent> handler = (c, e) -> CreateStorageAccountAction.createStorageAccount((e.getProject()));
+        am.registerHandler(ResourceCommonActionsContributor.CREATE, condition, handler);
+    }
+}
