+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-storage/src/main/java/com/microsoft/azure/toolkit/intellij/storage/component/KindComboBox.java
@@ -10,18 +10,16 @@ import com.microsoft.azure.toolkit.lib.Azure;
 import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
 import com.microsoft.azure.toolkit.lib.storage.model.Kind;
 import com.microsoft.azure.toolkit.lib.storage.model.Performance;
-import com.microsoft.azure.toolkit.lib.storage.model.Redundancy;
 import com.microsoft.azure.toolkit.lib.storage.service.AzureStorageAccount;
-import com.microsoft.azuretools.azurecommons.helpers.NotNull;
 
+import javax.annotation.Nonnull;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 
-public class RedundancyComboBox extends AzureComboBox<Redundancy> {
+public class KindComboBox extends AzureComboBox<Kind> {
 
     private Performance performance;
-    private Kind kind;
 
     public void setPerformance(Performance performance) {
         if (Objects.equals(performance, this.performance)) {
@@ -35,31 +33,18 @@ public class RedundancyComboBox extends AzureComboBox<Redundancy> {
         this.refreshItems();
     }
 
-    public void setKind(Kind kind) {
-        if (Objects.equals(kind, this.kind)) {
-            return;
-        }
-        this.kind = kind;
-        if (kind == null) {
-            this.clear();
-            return;
-        }
-        this.refreshItems();
-    }
-
-    @NotNull
+    @Nonnull
     @Override
     @AzureOperation(
-            name = "storage|account.redundancy.list.supported",
+            name = "storage|account.kind.list.supported",
             type = AzureOperation.Type.SERVICE
     )
-    protected List<? extends Redundancy> loadItems() {
-        return Objects.isNull(this.performance) ? Collections.emptyList() :
-                Azure.az(AzureStorageAccount.class).listSupportedRedundancies(this.performance, this.kind);
+    protected List<? extends Kind> loadItems() {
+        return Objects.isNull(this.performance) ? Collections.emptyList() : Azure.az(AzureStorageAccount.class).listSupportedKinds(this.performance);
     }
 
     @Override
     protected String getItemText(Object item) {
-        return item instanceof Redundancy ? ((Redundancy) item).getLabel() : super.getItemText(item);
+        return item instanceof Kind ? ((Kind) item).getLabel() : super.getItemText(item);
     }
 }
