+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-storage/src/main/java/com/microsoft/azure/toolkit/intellij/storage/creation/VMStorageAccountCreationDialog.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
-package com.microsoft.azure.toolkit.intellij.storage.component;
+package com.microsoft.azure.toolkit.intellij.storage.creation;
 
 import com.intellij.openapi.project.Project;
 import com.microsoft.azure.toolkit.lib.storage.model.Kind;
@@ -25,5 +25,4 @@ public class VMStorageAccountCreationDialog extends BaseStorageAccountCreationDi
         kindComboBox.setItemsLoader(() -> Arrays.asList(Kind.STORAGE_V2, Kind.STORAGE));
         redundancyComboBox.setItemsLoader(() -> Arrays.asList(Redundancy.PREMIUM_LRS, Redundancy.STANDARD_GRS, Redundancy.STANDARD_RAGRS));
     }
-
 }
