+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-storage/src/main/resources/META-INF/azure-intellij-plugin-storage.xml
@@ -1,0 +1,7 @@
+<idea-plugin>
+    <extensions defaultExtensionNs="com.microsoft.tooling.msservices.intellij.azure">
+        <azureExplorerContributor implementation="com.microsoft.azure.toolkit.ide.storage.StorageExplorerContributor"/>
+        <actions implementation="com.microsoft.azure.toolkit.ide.storage.StorageActionsContributor"/>
+        <actions implementation="com.microsoft.azure.toolkit.intellij.storage.IntellijStorageActionsContributor"/>
+    </extensions>
+</idea-plugin>
