+++ PluginsAndFeatures/azure-toolkit-for-intellij/settings.gradle
@@ -7,3 +7,4 @@ include 'azure-intellij-resource-connector-aad'
 include 'azure-intellij-plugin-springcloud'
 include 'azure-intellij-plugin-service-explorer'
 include 'azure-intellij-plugin-vm'
+include 'azure-intellij-plugin-storage'
